fix: update q

+++ agents/erode.py
@@ -295,7 +295,7 @@ class Agent(Base):
         self.Q2.optimizer.zero_grad()
         value_loss = 0
         for t in range(self.cfg.horizon):
-            Q1, Q2 = self.Q(zs[:, t, :], act_trajs[:, t, :])
+            Q1, Q2 = self.Q(zs[:, t, :], act_trajs[:, t, :].float())
             z_, reward = zs_[:, t, :], reward_trajs[:, t, :]
             td_target = self.td_target(z_, reward)
